requirements.txt
django
psycopg2-binary
python-dotenv
pillow
crispy-bootstrap5
gunicorn 
dj-database-url 
psycopg2-binary
django-allauth[mfa]
django-phonenumber-field[phonenumbers]